+++ src/pallet/actions_impl.clj
@@ -5,7 +5,8 @@
    [clojure.string :as string]
    [pallet.context :as context]
    [pallet.script.lib :as lib]
-   [pallet.stevedore :as stevedore])
+   [pallet.stevedore :as stevedore]
+   [pallet.utils :refer [base64-md5]])
   (:use
    [pallet.action :only [defaction]]
    [pallet.common.context :only [throw-map]]
@@ -17,7 +18,7 @@
   [name & script]
   `(stevedore/checked-script
     (if-let [context# (seq (context/phase-contexts))]
-      (str (string/join ": " context#) "\n" ~name)
+      (str (string/join ": " context#) " " ~name)
       ~name)
     ~@script))
 
@@ -87,3 +88,8 @@ to deal with local file transfer."
 (defmethod service-script-path :upstart
   [_ service-name]
   (str (stevedore/script (~lib/upstart-script-dir)) "/" service-name ".conf"))
+
+(defn temp-filename
+  "Generate a temporary file name for a given path."
+  [path]
+  (str (base64-md5 path) "-content"))