+++ src/pallet/project/loader.clj
@@ -17,17 +17,12 @@
       (update-in [:phases :bootstrap]
                  #(or % (plan-fn
                           (package-manager :update)
-                          (automated-admin-user))))
-      ;; (update-in [:phases :configure]
-      ;;            #(or % (plan-fn)))
-      ;; (update-in [:phases :settings]
-      ;;            #(or % (plan-fn)))
-      ))
+                          (automated-admin-user))))))
 
 (defn add-default-group-spec
   [{:keys [project-name phases] :as project}]
   (update-in project [:groups] #(or %
-                                    [(group-spec (keyword project-name)
+                                    [(group-spec project-name
                                                  :phases phases
                                                  :count 1)])))