+++ test/pallet/compute/implementation_test.clj
@@ -4,4 +4,4 @@
    pallet.compute.implementation))
 
 (deftest supported-providers-test
-  (is (= #{"node-list" "hybrid"} (set (supported-providers)))))
+  (is (= #{"node-list" "localhost" "hybrid"} (set (supported-providers)))))